slaskdjlaskdjalsdkjadl

+++ backend/recipe/models.py
@@ -157,7 +157,7 @@ class Recipe(models.Model):
         Tag,
         verbose_name='Теги',
     )
-    cooking_time = models.PositiveSmallIntegerField(
+    cooking_time = models.PositiveIntegerField(
         validators=(
             validators.MinValueValidator(
                 MIN_COOKING_TIME
